pyproject.toml
[tool.poetry]
name = "educa"
version = "0.1.0"
description = ""
authors = ["Gabriel Lustosa <lustosaki2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
django-ninja = "^0.22.1"
django-ordered-model = "^3.7.4"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
factory-boy = "^3.2.1"
pytube = "^15.0.0"
pillow = "^9.5.0"
django-mptt = "^0.14.0"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.5.2"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
blue = "^0.9.1"
isort = "^5.12.0"
taskipy = "^1.10.4"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.1.9"
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "educa.settings.test"
addopts = "--doctest-modules --strict-markers"


[tool.isort]
profile = "black"
line_length = 79


[tool.taskipy.tasks]
lint = "blue . && isort ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x"
coverage = "coverage html"